Fix bug for when outputs are added to CopyTable.
When performing re-write (optimsiation) passes over the logical
plan/graph, a CopyTable is used to ensure that if a node has already
been re-written that it is not re-written again when it is traversed to
from another output (top-level) node. Specific re-writes are responsible
for adding nodes to the CopyTable but sometimes they do not for the
purpoe of replicating certain nodes (e.g. Flatten and GroupByKey
splitting). Because the specific re-writes are already responsible for
adding nodes to the CopyTable, it is not necessary, and incorrect, for
the top-level traversal (travOnce) to add output (top-level) nodes. This
bug fix makes a change in travOnce so that output nodes are not added
to the CopyTable.