Fix bug in generation of identity reducers.
Identity reducers are required for boty Bypass Output Channels as well
as Flatten Output Channels. In the case of the former, the input will be
from either a Mapper or GbkMapper. Depending on which, the generation of
an identity recucer will be different.

Fix a bug where the identity reducer implementation assumed the input
was from a Mapper. The solution only makes the assumption that the input
is either a Mapper or GbkMapper by making them both ReducerLike and
implementing mkTaggedReducer for each. Similarly, Flatten has
ReducerLike mixed-in for its implementation of an identity reducer in
the case of Flatten Output Channels.